+++ lib/jessie.dart
@@ -1,5 +1,0 @@
-/// JSONPath for Dart
-library jessie;
-
-export 'package:jessie/src/json_path.dart';
-export 'package:jessie/src/result.dart';