+++ lib/json_path.dart
@@ -1,0 +1,5 @@
+/// JSONPath for Dart
+library json_path;
+
+export 'package:json_path/src/json_path.dart';
+export 'package:json_path/src/result.dart';