+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## [Unreleased]
+## [0.0.0+dev.2] - 2020-07-28
+### Added
+- Recursive selector
+- Wildcard selector
+
 ## [0.0.0+dev.1] - 2020-07-27
 ### Added
 - Tokenizer and AST
@@ -8,5 +13,6 @@
 ### Added
 - Basic design draft
 
-[Unreleased]: https://github.com/f3ath/jessie/compare/0.0.0+dev.1...HEAD
+[Unreleased]: https://github.com/f3ath/jessie/compare/0.0.0+dev.2...HEAD
+[0.0.0+dev.2]: https://github.com/f3ath/jessie/compare/0.0.0+dev.1...0.0.0+dev.2
 [0.0.0+dev.1]: https://github.com/f3ath/jessie/compare/0.0.0+dev.0...0.0.0+dev.1